add first unit test for processes
