+++ frontend/package.json
@@ -20,6 +20,7 @@
     "astro": "^2.3.2",
     "babel-jest": "^29.6.2",
     "jest": "^29.4.3",
+    "jest-environment-jsdom": "^29.6.2",
     "svelte": "^3.50.1",
     "svelte-jester": "^3.0.0",
     "svelte-switch-case": "^2.0.0"
